+++ releaser/internal/common/test_data/verify_git_tags_do_not_already_exist/versions_invalid_syntax.yaml
@@ -13,17 +13,10 @@
 # limitations under the License.
 
 module-sets:
-  mod-set-1:
-    version: v1.2.3-RC1+meta
+  - should-not-have-front-dash:
+    version: v1.0.0
     modules:
       - go.opentelemetry.io/test/test1
-  mod-set-2:
-    version: v0.1.0
-    modules:
-      - go.opentelemetry.io/test3
-  mod-set-3:
-    version: v2.2.2
-    modules:
-      - go.opentelemetry.io/testroot/v2
+      - go.opentelemetry.io/test/test2
 excluded-modules:
-  - go.opentelemetry.io/test/testexcluded
+  - go.opentelemetry.io/excluded1
