+++ releaser/internal/common/versions.go
@@ -20,8 +20,6 @@ import (
 	"io/ioutil"
 	"path/filepath"
 
-	"github.com/go-git/go-git/v5"
-	"github.com/go-git/go-git/v5/plumbing"
 	"github.com/spf13/viper"
 	"golang.org/x/mod/modfile"
 )
@@ -30,157 +28,6 @@ const (
 	repoRootTag = ModuleTagName("REPOROOTTAG")
 	SemverRegex = `\s+v(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?`
 )
-
-// ModuleVersioning holds info about modules listed in a versioning file.
-type ModuleVersioning struct {
-	ModSetMap  ModuleSetMap
-	ModPathMap ModulePathMap
-	ModInfoMap ModuleInfoMap
-}
-
-// NewModuleVersioning returns a ModuleVersioning struct from a versioning file and repo root.
-func NewModuleVersioning(versioningFilename string, repoRoot string) (ModuleVersioning, error) {
-	vCfg, err := readVersioningFile(versioningFilename)
-	if err != nil {
-		return ModuleVersioning{}, fmt.Errorf("error reading versioning file %v: %v", versioningFilename, err)
-	}
-
-	modSetMap, err := vCfg.buildModuleSetsMap()
-	if err != nil {
-		return ModuleVersioning{}, fmt.Errorf("error building module set map for NewModuleVersioning: %v", err)
-	}
-
-	modInfoMap, err := vCfg.buildModuleMap()
-	if err != nil {
-		return ModuleVersioning{}, fmt.Errorf("error building module info map for NewModuleVersioning: %v", err)
-	}
-
-	modPathMap, err := vCfg.BuildModulePathMap(repoRoot)
-	if err != nil {
-		return ModuleVersioning{}, fmt.Errorf("error building module path map for NewModuleVersioning: %v", err)
-	}
-
-	return ModuleVersioning{
-		ModSetMap:  modSetMap,
-		ModPathMap: modPathMap,
-		ModInfoMap: modInfoMap,
-	}, nil
-}
-
-// ModuleSetRelease contains info about a specific set of modules in the versioning file to be updated.
-type ModuleSetRelease struct {
-	ModuleVersioning
-	ModSetName string
-	ModSet     ModuleSet
-	TagNames   []ModuleTagName
-	Repo       *git.Repository
-}
-
-// NewModuleSetRelease returns a ModuleSetRelease struct by specifying a specific set of modules to update.
-func NewModuleSetRelease(versioningFilename, modSetToUpdate, repoRoot string) (ModuleSetRelease, error) {
-	modVersioning, err := NewModuleVersioning(versioningFilename, repoRoot)
-	if err != nil {
-		return ModuleSetRelease{}, fmt.Errorf("unable to load baseVersionStruct: %v", err)
-	}
-
-	// get new version and mod tags to update
-	modSet, exists := modVersioning.ModSetMap[modSetToUpdate]
-	if !exists {
-		return ModuleSetRelease{}, fmt.Errorf("could not find module set %v in versioning file", modSetToUpdate)
-	}
-
-	// get tag names of mods to update
-	tagNames, err := ModulePathsToTagNames(
-		modSet.Modules,
-		modVersioning.ModPathMap,
-		repoRoot,
-	)
-	if err != nil {
-		return ModuleSetRelease{}, fmt.Errorf("could not retrieve tag names from module paths: %v", err)
-	}
-
-	repo, err := git.PlainOpen(repoRoot)
-	if err != nil {
-		return ModuleSetRelease{}, fmt.Errorf("error getting git.Repository from repo root dir %v: %v", repoRoot, err)
-	}
-
-	return ModuleSetRelease{
-		ModuleVersioning: modVersioning,
-		ModSetName:       modSetToUpdate,
-		ModSet:           modSet,
-		TagNames:         tagNames,
-		Repo:             repo,
-	}, nil
-
-}
-
-// ModSetVersion gets the version of the module set to update.
-func (modRelease ModuleSetRelease) ModSetVersion() string {
-	return modRelease.ModSet.Version
-}
-
-// ModSetPaths gets the import paths of all modules in the module set to update.
-func (modRelease ModuleSetRelease) ModSetPaths() []ModulePath {
-	return modRelease.ModSet.Modules
-}
-
-// ModSetTagNames gets the tag names of all modules in the module set to update.
-func (modRelease ModuleSetRelease) ModSetTagNames() []ModuleTagName {
-	return modRelease.TagNames
-}
-
-// ModuleFullTagNames gets the full tag names (including the version) of all modules in the module set to update.
-func (modRelease ModuleSetRelease) ModuleFullTagNames() []string {
-	return combineModuleTagNamesAndVersion(modRelease.ModSetTagNames(), modRelease.ModSetVersion())
-}
-
-// VerifyGitTagsDoNotAlreadyExist checks if Git tags have already been created that match the specific module tag name
-// and version number for the modules being updated. If the tag already exists, an error is returned.
-func (modRelease ModuleSetRelease) VerifyGitTagsDoNotAlreadyExist() error {
-	newTags := make(map[string]bool)
-
-	modFullTags := modRelease.ModuleFullTagNames()
-
-	for _, newFullTag := range modFullTags {
-		newTags[newFullTag] = true
-	}
-
-	existingTags, err := modRelease.Repo.Tags()
-	if err != nil {
-		return fmt.Errorf("error getting repo tags: %v", err)
-	}
-
-	var existingGitTagNames []string
-
-	err = existingTags.ForEach(func(ref *plumbing.Reference) error {
-		tagObj, err := modRelease.Repo.TagObject(ref.Hash())
-		if err != nil {
-			return fmt.Errorf("error retrieving tag object: %v", err)
-		}
-		if _, exists := newTags[tagObj.Name]; exists {
-			existingGitTagNames = append(existingGitTagNames, tagObj.Name)
-		}
-
-		return nil
-	})
-	if err != nil {
-		return fmt.Errorf("could not check all git tags: %v", err)
-	}
-
-	if len(existingGitTagNames) > 0 {
-		return &ErrGitTagsAlreadyExists{
-			tagNames: existingGitTagNames,
-		}
-	}
-
-	return nil
-}
-
-// versionConfig is needed to parse the versions.yaml file with viper.
-type versionConfig struct {
-	ModuleSets      ModuleSetMap `mapstructure:"module-sets"`
-	ExcludedModules []ModulePath `mapstructure:"excluded-modules"`
-}
 
 // excludedModules functions as a set containing all module paths that are excluded
 // from versioning.
@@ -218,6 +65,12 @@ type ModulePathMap map[ModulePath]ModuleFilePath
 // ModuleTagName is the simple file path to the directory of a go.mod file used for Git tagging.
 // For example, the opentelemetry-go/sdk/metric/go.mod file will have a ModuleTagName "sdk/metric".
 type ModuleTagName string
+
+// versionConfig is needed to parse the versions.yaml file with viper.
+type versionConfig struct {
+	ModuleSets      ModuleSetMap `mapstructure:"module-sets"`
+	ExcludedModules []ModulePath `mapstructure:"excluded-modules"`
+}
 
 // readVersioningFile reads in a versioning file (typically given as versions.yaml) and returns
 // a versionConfig struct.
